requirements.txt
praw
confluent-kafka
pymongo
apache-airflow-providers-apache-kafka
faiss-cpu
sentence-transformers
matplotlib
pyspark
bs4
lxml
html5lib


